// frontend/src/api/model/githubCom4H1RZooraInternalDomainConversationMember.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */
import type { GithubCom4H1RZooraInternalDomainConversationMemberRole } from './githubCom4H1RZooraInternalDomainConversationMemberRole';
import type { GithubCom4H1RZooraInternalDomainUser } from './githubCom4H1RZooraInternalDomainUser';

export interface GithubCom4H1RZooraInternalDomainConversationMember {
  conversation_id?: string;
  id?: string;
  joined_at?: string;
  last_read_at?: string;
  last_read_message_id?: string;
  muted_until?: string;
  role?: GithubCom4H1RZooraInternalDomainConversationMemberRole;
  user?: GithubCom4H1RZooraInternalDomainUser;
  user_id?: string;
}